# Conflicts:
#   frontend/src/pages/teacher/TeacherBroadcast.tsx

Reviewer: 김지원, 김범수

+++ backend/src/main/java/com/ssafy/kidslink/application/schedule/repository/TeacherScheduleRepository.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface TeacherScheduleRepository extends JpaRepository<TeacherSchedule, Long> {
+public interface TeacherScheduleRepository extends JpaRepository<TeacherSchedule, Integer> {
     List<TeacherSchedule> findByTeacherAndTeacherScheduleDate(Teacher teacher, LocalDate date);
 }
